Keep validate arithmetic in PLN domain, use PLN in IdentityError
Replace Double with PLN in identity table and IdentityError fields.
toDouble conversion only at system boundaries (stderr formatting,
test assertions).

+++ src/test/scala/sfc/accounting/BopSfcSpec.scala
@@ -7,7 +7,7 @@ import sfc.types.*
 class BopSfcSpec extends AnyFlatSpec with Matchers:
 
   private def errorDelta(result: Either[Vector[Sfc.IdentityError], Unit], id: Sfc.SfcIdentity): Double =
-    result.swap.getOrElse(Vector.empty).find(_.identity == id).map(e => e.actual - e.expected).getOrElse(0.0)
+    result.swap.getOrElse(Vector.empty).find(_.identity == id).map(e => (e.actual - e.expected).toDouble).getOrElse(0.0)
 
   private def zeroSnap: Sfc.Snapshot = Sfc.Snapshot(
     hhSavings = PLN.Zero, hhDebt = PLN.Zero, firmCash = PLN.Zero, firmDebt = PLN.Zero,
