+++ src/test/scala/sfc/accounting/SfcPropertySpec.scala
@@ -13,7 +13,7 @@ class SfcPropertySpec extends AnyFlatSpec with Matchers with ScalaCheckPropertyC
     PropertyCheckConfiguration(minSuccessful = 200)
 
   private def errorDelta(result: Either[Vector[Sfc.IdentityError], Unit], id: Sfc.SfcIdentity): Double =
-    result.swap.getOrElse(Vector.empty).find(_.identity == id).map(e => e.actual - e.expected).getOrElse(0.0)
+    result.swap.getOrElse(Vector.empty).find(_.identity == id).map(e => (e.actual - e.expected).toDouble).getOrElse(0.0)
 
   // --- Consistent flows always pass ---
